requirements.txt
# Core Security Libraries
requests==2.26.0
python-nmap==0.6.1
scapy==2.4.5
shodan==1.25.0

# Machine Learning and AI
numpy==1.21.2
tensorflow==2.6.0
scikit-learn==0.24.2
torch==1.9.1
keras==2.6.0

# Cryptography
cryptography==3.4.8
pynacl==1.4.0

# Network and DNS
dnspython==2.1.0
whois==0.9.13

# Visualization
plotly==5.3.1
matplotlib==3.4.3
seaborn==0.11.2

# System and Performance
psutil==5.8.0
speedtest-cli==2.1.3

# Logging and Monitoring
rich==10.12.0
colorama==0.4.4

# Development and Testing
pytest==6.2.5
mypy==0.910
black==21.9b0

# Optional Advanced Tools
python-dotenv
memory_profiler

# Web Technologies
beautifulsoup4==4.10.0
selenium==4.0.0
